feat: create bundle size module
