+++ packages/jest-puppeteer-env/src/performance/index.js
@@ -1,4 +1,4 @@
-const getBytes = require('bytes');
+const { checkBundleSize: checkBundleSizeFactory } = require('./bundleSize');
 
 const performance = async ({ page, client, config }) => {
   const enabling = client.send('Performance.enable');
@@ -14,38 +14,9 @@ const performance = async ({ page, client, config }) => {
       pageMetrics
     }));
 
-  const checkBundleSize = (response) => {
-    let bundleSizeViolation;
-    const headers = response.headers();
-    const contentLength = Number(headers['content-length']);
-    const contentType = String(headers['content-type']);
-    const isJavaScript = contentType.includes('javascript');
-    const isStylesheet = contentType.includes('css');
-    const url = response.url();
-    const bundleSizes = config.performance?.bundleSizes;
-    const matchingPatterns = bundleSizes.filter(({ path }) => new RegExp(path, 'u').test(url));
-
-    if (matchingPatterns.length > 1) {
-      throw new Error(`URL matches multiple bundle size paths: ${matchingPatterns.join(', ')}`);
-    }
-
-    const shouldCheckBundleSize = matchingPatterns.length && (isJavaScript || isStylesheet) && contentLength;
-
-    if (shouldCheckBundleSize) {
-      const maxBundleSize = matchingPatterns[0].maxSize;
-      const bytes = getBytes(maxBundleSize);
-
-      if (contentLength > bytes) {
-        bundleSizeViolation = {
-          rule: matchingPatterns[0],
-          size: getBytes(contentLength),
-          url
-        };
-      }
-    }
-
-    return bundleSizeViolation;
-  };
+  const checkBundleSize = config.performance?.bundleSizes
+    ? checkBundleSizeFactory({ bundleSizes: config.performance.bundleSizes })
+    : undefined;
 
   return {
     checkBundleSize,
